sepparated search from crawler logic
